moved redeem stake test
